+++ Providers/DieProvider.cs
@@ -15,7 +15,7 @@ namespace VueExample.Providers
         {
             _srv6Context = srv6Context;
         }
-        public async Task<DieParameterOld> GetOrAddDieParameter(long dieId, int measurementRecordingId, int parameterId = 247, string value = "0")
+        public async Task<DieParameterOld> GetOrAddDieParameter(long dieId, int measurementRecordingId, short parameterId = 247, string value = "0")
         {
             var dieParameter = await _srv6Context.DiesParameterOld.FirstOrDefaultAsync(x => x.MeasurementRecordingId == measurementRecordingId && x.DieId == dieId && x.Id247 == parameterId);
             if(dieParameter is null)
